update namespace, rename class
